+++ index.html
@@ -8,14 +8,14 @@
 </head>
 <body>
 	<input id="dtp" type="datetime" name="date" value="2017-03-28 17:13:50">
-	<script src="index.js"></script>
 
-	<script>
-	(function() {
-		var element = document.getElementById('dtp');
-		var ob = new MD(element);
-		ob.view();
-	})();
+	<script type="module">
+		import MD from "./src/MD.js";
+		(function() {
+			var element = document.getElementById('dtp');
+			var ob = new MD(element);
+			ob.view();
+		})();
 	</script>
 </body>
 </html>
